python-workers/requirements.txt
fastapi
uvicorn
celery
pdfplumber
pytesseract
ffmpeg-python
requests
python-dotenv
redis
qdrant-client
